+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/client/AltColorServiceClient.java
@@ -2,6 +2,7 @@ package de.mephisto.vpin.restclient.client;
 
 import de.mephisto.vpin.restclient.*;
 import de.mephisto.vpin.restclient.jobs.JobExecutionResult;
+import de.mephisto.vpin.restclient.representations.altcolor.AltColor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpMethod;