+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/representations/altsound/AltSound.java
@@ -1,4 +1,4 @@
-package de.mephisto.vpin.restclient;
+package de.mephisto.vpin.restclient.representations.altsound;
 
 import de.mephisto.vpin.restclient.representations.ValidationState;