+++ vpin-studio-rest-client/src/main/java/module-info.java
@@ -14,4 +14,7 @@ open module de.mephisto.vpin.restclient {
   exports de.mephisto.vpin.restclient.jobs;
   exports de.mephisto.vpin.restclient.popper;
   exports de.mephisto.vpin.restclient.mame;
+  exports de.mephisto.vpin.restclient.representations.altcolor;
+  exports de.mephisto.vpin.restclient.representations.altsound;
+  exports de.mephisto.vpin.restclient.representations.vpx;
 }