+++ vpin-studio-server/src/test/java/de/mephisto/vpin/server/vpx/VPXUtilTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 
 import java.io.File;
 import java.util.Map;
+import java.util.Set;
 
 import static de.mephisto.vpin.server.AbstractVPinServerTest.TABLE_NAMES;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -25,6 +26,11 @@ public class VPXUtilTest {
     for (String tableName : TABLE_NAMES) {
       File table = new File("../testsystem/vPinball/VisualPinball/Tables/" + tableName);
       Map<String, String> data = VPXUtil.readTableInfo(table);
+      Set<Map.Entry<String, String>> entries = data.entrySet();
+      for (Map.Entry<String, String> entry : entries) {
+        System.out.println(entry.getKey() + " => " + entry.getValue());
+      }
+
       assertNotNull(data);
     }
   }